update ad api
